Revert "Merge pull request #1252 from lykkin/revert-logstash-output"
This reverts commit 9b0f8a1c5ba16a5ff799e90574a608d759882f4a, reversing
changes made to 330735c728e9e56ca05e4e384c7da006295b174a.
