+++ cmd/fleet/error.go
@@ -26,12 +26,11 @@ const (
 	EcsHttpResponseCode      = logger.EcsHttpResponseCode
 	EcsHttpResponseBodyBytes = logger.EcsHttpResponseBodyBytes
 
-	LogApiKeyId              = logger.ApiKeyId
-	LogPolicyId              = logger.PolicyId
-	LogAgentId               = logger.AgentId
-	LogEnrollApiKeyId        = logger.EnrollApiKeyId
-	LogAccessApiKeyId        = logger.AccessApiKeyId
-	LogDefaultOutputApiKeyId = logger.DefaultOutputApiKeyId
+	LogAPIKeyID       = logger.ApiKeyId
+	LogPolicyID       = logger.PolicyId
+	LogAgentID        = logger.AgentId
+	LogEnrollAPIKeyID = logger.EnrollApiKeyId
+	LogAccessAPIKeyID = logger.AccessApiKeyId
 )
 
 type errResp struct {
